+++ tests/test_checker/test_noqa.py
@@ -168,6 +168,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS444': 1,
     'WPS445': 1,
     'WPS446': 1,
+    'WPS447': 1,
 
     'WPS500': 1,
     'WPS501': 1,