+++ CHANGELOG.md
@@ -11,6 +11,7 @@ We used to have incremental versioning before `0.1.0`.
 - `WPS431` now allow customize whitelist via `nested-classes-whitelist` setting
 - Forbids to have invalid strings like `**{'@': 1}`
 - Forbids to use implicit primitive values in a form of `lambda`
+- Forbids to use approximate math constants
 
 
 ## 0.12.5